+++ source/GameAction.h
@@ -1,5 +1,5 @@
-/* MissionAction.h
-Copyright (c) 2014 by Michael Zahniser
+/* GameAction.h
+Copyright (c) 2020 by Michael Zahniser
 
 Endless Sky is free software: you can redistribute it and/or modify it under the
 terms of the GNU General Public License as published by the Free Software
@@ -10,12 +10,11 @@ WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
 PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 */
 
-#ifndef MISSION_ACTION_H_
-#define MISSION_ACTION_H_
+#ifndef GAME_ACTION_H_
+#define GAME_ACTION_H_
 
 #include "ConditionSet.h"
 #include "Conversation.h"
-#include "GameAction.h"
 #include "LocationFilter.h"
 #include "Phrase.h"
 
@@ -27,7 +26,6 @@ PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 
 class DataNode;
 class DataWriter;
-class GameAction;
 class GameEvent;
 class Outfit;
 class PlayerInfo;
@@ -41,53 +39,31 @@ class UI;
 // milestone: offered, accepted, declined, completed or failed. Actions might
 // include showing a dialog or conversation, giving the player payment or a
 // special item, modifying condition flags, or queueing an event to occur.
-class MissionAction {
+class GameAction {
 public:
-	MissionAction() = default;
+	GameAction() = default;
 	// Construct and Load() at the same time.
-	MissionAction(const DataNode &node, const std::string &missionName);
+	GameAction(const DataNode &node, const std::string &missionName);
 	
 	void Load(const DataNode &node, const std::string &missionName);
-	// Note: the Save() function can assume this is an instantiated mission, not
-	// a template, so it only has to save a subset of the data.
+	void LoadSingle(const DataNode &child, const std::string &missionName);
 	void Save(DataWriter &out) const;
 	
 	bool IsEmpty() const;
 	
 	int Payment() const;
 	
-	const std::string &DialogText() const;
-	
-	// Check if this action can be completed right now. It cannot be completed
-	// if it takes away money or outfits that the player does not have, or should
-	// take place in a system that does not match the specified LocationFilter.
-	bool CanBeDone(const PlayerInfo &player, const std::shared_ptr<Ship> &boardingShip = nullptr) const;
-	// Perform this action. If a conversation is shown, the given destination
-	// will be highlighted in the map if you bring it up.
-	void Do(PlayerInfo &player, UI *ui = nullptr, const System *destination = nullptr, const std::shared_ptr<Ship> &ship = nullptr, const bool isUnique = true) const;
+	// Perform this action.
+	void Do(PlayerInfo &player, bool conversation = false) const;
 	
 	// "Instantiate" this action by filling in the wildcard text for the actual
 	// destination, payment, cargo, etc.
-	MissionAction Instantiate(std::map<std::string, std::string> &subs, const System *origin, int jumps, int payload) const;
+	GameAction Instantiate(std::map<std::string, std::string> &subs, int jumps, int payload) const;
 	
 	
 private:
-	std::string trigger;
-	std::string system;
-	LocationFilter systemFilter;
+	bool empty = true;
 	
-	std::string dialogText;
-	const Phrase *stockDialogPhrase = nullptr;
-	Phrase dialogPhrase;
-	
-	const Conversation *stockConversation = nullptr;
-	Conversation conversation;
-	
-	std::map<const Outfit *, int> gifts;
-	std::map<const Outfit *, int> requiredOutfits;
-	
-	GameAction actions;
-	/* Now handled by GameAction
 	std::string logText;
 	std::map<std::string, std::map<std::string, std::string>> specialLogText;
 	
@@ -99,7 +75,6 @@ private:
 	std::set<std::string> fail;
 	
 	ConditionSet conditions;
-	*/
 };
